[dev]: multipart upload test

+++ backend/FileService/FileService.Contracts/Responses/StartMultipartUploadFileResponse.cs
@@ -4,5 +4,5 @@ public record StartMultipartUploadFileResponse(
     Guid MediaAssetId,
     string UploadId,
     IReadOnlyList<ChunkUploadUrl> ChunkUploadUrls,
-    long ChunkSize
+    int ChunkSize
     );
